requirements.txt
squareconnect==2.20180918.0
numpy==1.15.2
pandas==0.23.4
psycopg2==2.7.5
SQLAlchemy==1.2.12
PyYAML==3.13
requests==2.19.1